feat(onboarding): 首次引导在大厅逐步高亮真实入口，并展示示例项目卡
首弹引导扩展为大厅四步：欢迎、新建项目入口、示例项目卡、设置入口与未配置红点，
三语文案齐备。示例卡复用大厅真实卡片的只读形态，仅在引导运行期间展示，不产生任何
项目数据。锚点集中登记为联合类型，界面与步骤两头引用同一份注册表，名字漂移在编译期
暴露；锚点找不到时该步降级为居中气泡，引导继续。
