+++ frontend/src/i18n/en/onboarding.ts
@@ -3,8 +3,12 @@ export default {
   // Tour steps
   'welcome_title': 'Welcome to [[brand]]',
   'welcome_body': 'Hand it a novel and it breaks the story into shots, generates the frames, and cuts them into a short video. This tour only walks through the interface — it changes none of your data.',
-  'finish_title': 'Your turn',
-  'finish_body': 'Start by importing a novel, then take the rest one step at a time. To see this again, open Settings → About.',
+  'lobby_create_title': 'Start a project here',
+  'lobby_create_body': 'Every project starts from a novel. Import a .txt, .docx, .epub, or .pdf and [[brand]] reads it into episodes you can work through.',
+  'lobby_demo_title': 'A project once it gets going',
+  'lobby_demo_body': 'This card is a sample, not one of your projects. The badge shows the current phase, and the counters below track characters, scenes, props, and episodes as production fills them in.',
+  'lobby_settings_title': 'Providers live in Settings',
+  'lobby_settings_body': 'Image, video, and text generation each run on a provider you choose. A red dot on this button means something required is still missing — open Settings and it points you at the gap.',
 
   // Tour controls
   'next': 'Continue',
@@ -13,6 +17,12 @@ export default {
   'skip': 'Skip',
   'close': 'Close the tour',
   'progress': 'Step {{current}} of {{total}}',
+
+  // Demo card shown during the tour
+  'demo_section_eyebrow': 'Sample project',
+  'demo_section_note': 'Shown during the tour only',
+  'demo_project_title': "Alice's Adventures in Wonderland",
+  'demo_project_style': 'Watercolour storybook',
 
   // Settings → About entry
   'replay_title': 'Product tour',
